Update Coding project.py
w - OMGOMGOMGOMG I think I've cracked it. The blocks still don't move down fully but the main problem has been found. I had been using self.column and self.row without actually initializing it properly according to its position in store block. THAT IS WHY IT SPAWNS IN THE LEFT HAND CORNER BECAUSE THE VALUES OF COLUMN AND ROW ARE 0, 0 FROM WHEN IT IS FIRST CALLED!