Update Coding project.py
w - all of the blocks move down properly now. Still need to fix the edge boundary value as blocks stop 1 block above bottom and the blocks still spawn on each other. Also, some of the blocks continue to fall even though it had moved as low as it could go.