Update Coding project.py
w - [Pic: switchscreenglitch] Can only select one screen change and the program hangs (becomes unresponsive) when trying to move onto the third screen, the second screen change. The issue is resolved in the next page because of initialization of donestartscreen etc within the donegame loop since no loop would be able to run and pass the entry condition because once donestartscreen is changed is True (to break out of loop), it's then unchanged.